Move AbstratView to root directory. Change view import modules. Code review
