+++ src/view/controls/controls.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract/abstract';
+import AbstractView from '../abstract';
 
 const createControlsTemplate = () => {
   return (
