+++ src/view/days/days.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract/abstract';
+import AbstractView from '../abstract';
 
 const createDaysTemplate = () => {
   return (
