+++ src/view/points-list/points-list.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract/abstract';
+import AbstractView from '../abstract';
 
 const createPointsListTemplate = () => {
   return (
