+++ src/view/statistics/statistics.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract/abstract';
+import AbstractView from '../abstract';
 
 const createStatisticsTemplate = () => {
   return (
